feat: Implement checkout components and hooks
- Add CheckoutEmptyState component for empty cart state.
- Create CheckoutHeader component to display checkout progress messages.
- Implement CheckoutLoadingState component for loading indication during checkout.
- Develop CheckoutProgress component to visualize the checkout steps.
- Add PaymentMethodCard component for selecting payment methods.
- Create ShippingMethodCard component for selecting shipping carriers and options.
- Implement SpecialInstructionsCard for additional delivery notes.
- Develop useCheckoutForm hook to manage checkout state and handle order placement.
